KeyHandler: Only restore mute media volume if stream is still 0
This is a better UX:
If we're sure the user have changed the volume while in this mode they probably don't want it to restore.
Doing this instead of setting up a volume change listener which would be too pricey for such a useless feature.
The switch summary already perfectly describes what it does.

Add some comments while at it to make it clear for future dev